Switch to using a ThemeProvider instead of a global module
By using a `styled-components` `<ThemeProvider />`, all styled
components have access to the global theme (previously called `Palette`)
via interpolating a function: `${props => props.theme}`.

This might seem like a tiny difference, but this means that we can now
dynamically update the theme of the interface. By using this, we get
Slack-like "Update your sidebar theme" FOR FREE! We can in the future let
frequencies specify a custom theme for their community, and it'll be
applied for all users. (+ instant preview without a page refresh or
hitting "save")

Fuck yeah.

I've refactored everything to use this pattern.