Create first sp.chat shortener for nicer meta tags
This redirects these two kinds of URLs:

- `/~frequency`
- `/story`

What it does differently though is that it adds two query params to the
URL it redirects to. Imagine a user clicks on this link: `/~travel`.

What happens is that we get the data for the frequency (in this case,
for the story in the other case) and attach is to the `t` and `d` query
params in the redirect. So that URL then redirects to:

```
https://spectrum.chat/~travel?t=travel&d=A%20community%20about%20going%20places.%0A%0A🛩️🛥️🚄🚶‍
```

From the main Spectrum codebase we can then grab those query params on
first render (!) and add nice meta tags. Related PR to main repo coming
soon.