Speed up getUserPermissionInCommunity
This function is called a lot and part of the reason why we've been so
slow recently. Instead of my dumbass `.between` (which was totally
unnecessary) we can just use `.getAll`, which speeds up the query ~2x
locally.

This might also enable using a loader for this query, which would be a
significant speedup. Digging into that now.