Fix subscriptions
When we switched to Redis for subscriptions, what happened was that
every subscription result was passed through Redis, which had to
serialize it, so we had to un-serialize it.

This is not only unnecessary right now, it also is (I think) the source
of the chat messages crashing sometimes when you send a message.