channelPendingUsers loader
For some reason this is fetched from the home community view for every
single channel and then never displayed, which is... suboptimal.
(even loads if you're not an admin, what)

Anyways, this implements a loader so it will be faster, but ideally we
shouldn't request this at all unless you go to the channel member view
and are a channel mod or something like that.