Fix message flicker when sending a message
Because both the mutation and the subscription for new messages in a
thread would add the new message to the state when the local user sends
one, the message would flicker sometimes because it was duplicated and
React got a key conflict.

This fixes it by ignoring duplicates in the subscription.