Refactor asyncify to only run one db listener
Previously we'd start one db listener for each user and subscription.
While we should do that eventually, it should be without listening to
changes of an entire table, instead only listening to relevant changes.

While that's possible, it's a bit too much work right now, so this patch
makes it so we only start a db listener once and then only give each
user a new queue.