Consistent linkification in messages
Previously, linkification in rendered (!) messages was totally different
from linkification in the chat input and the thread editor.

This makes them consistent, so if you type github.com/bla in a message
it gets rendered as a link in both the chat input and the rendered
bubble.

Note that this uses a list of TLDs under the hood, so things like
"spectrum.chat/figma" are linkified while "nota.urlextensions" won't be.