Cache homescreen icons from the ServiceWorker
We had to disable caching for all images as Zeit was rate-limiting us
due to too many requests being sent from a single browser in a short
period of time.

Unfortunately, that broke adding-to-homescreen as the icons wouldn't be
sent properly. This patch fixes it by caching homescreen icons in the
ServiceWorker. Closes #2822